Refactoring for clarity
