+++ asg-passes/src/dotify/mod.rs
@@ -14,967 +14,30 @@
 // You should have received a copy of the GNU General Public License
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 mod dotgraph;
+use dotgraph::*;
+
+mod expression;
+
+mod dotifier;
+pub use dotifier::Dotifier;
+
+mod statement;
+
+mod program;
 
 use leo_asg::*;
-use leo_errors::{Result, Span};
+use leo_errors::Result;
 
-use dot::{ArrowShape, Style};
-use dotgraph::{DotEdge, DotGraph, DotNode, LabelType};
-use petgraph::graph::NodeIndex;
-use std::collections::HashMap;
 use std::fs::File;
 use std::path::PathBuf;
 
-pub struct Dotify<'a, 'b> {
-    graph: DotGraph,
-    context: &'b AsgContext<'a>,
-    id_map: HashMap<u32, NodeIndex>,
-    edges: Vec<(u32, u32, (String, LabelType), Option<(String, LabelType)>)>, // For edges that are meant to be added after entire ASG is traversed
-}
-
-impl<'a, 'b> Dotify<'a, 'b> {
-    fn new(graph: DotGraph, context: &'b AsgContext<'a>) -> Self {
-        Dotify {
-            graph,
-            context,
-            id_map: HashMap::new(),
-            edges: Vec::new(),
-        }
-    }
-
-    // Helper functions to make it easier to construct graphs
-    fn add_or_get_node(&mut self, id: u32, label: String, typ: LabelType) -> NodeIndex {
-        let &mut Dotify {
-            ref mut id_map,
-            ref mut graph,
-            context: _,
-            edges: _,
-        } = self;
-        *id_map.entry(id).or_insert_with(|| {
-            let node = DotNode {
-                id: format!("N{:}", id),
-                shape: None,
-                label: (label, typ),
-                style: Style::None,
-                color: None,
-            };
-            graph.add_node(node)
-        })
-    }
-
-    fn add_edge(
-        &mut self,
-        start_idx: NodeIndex,
-        end_idx: NodeIndex,
-        label: (String, LabelType),
-        color: Option<(String, LabelType)>,
-    ) {
-        let edge = DotEdge {
-            start_idx,
-            end_idx,
-            label,
-            end_arrow: ArrowShape::NoArrow,
-            start_arrow: ArrowShape::NoArrow,
-            style: Style::None,
-            color,
-        };
-        self.graph.add_edge(edge);
-    }
-
-    fn add_remaining_edges(&mut self) {
-        // TODO make more idomatic
-        let mut edges = self.edges.drain(..).collect::<Vec<_>>();
-        for (start_id, end_id, label, color) in edges.drain(..) {
-            let start_idx = self.id_map.get(&start_id).unwrap(); // All nodes should have been added to ID map
-            let end_idx = self.id_map.get(&end_id).unwrap(); // All nodes should have been added to ID map
-            self.add_edge(*start_idx, *end_idx, label, color);
-        }
-    }
-
-    fn generate_span_info(span: &Option<Span>) -> String {
-        if let Some(span) = span {
-            format!(
-                "File: {:}\nLocation: {:}\n\nContent: {:}",
-                span.path.to_string(),
-                span,
-                span.content.to_string(),
-            )
-        } else {
-            "".to_string()
-        }
-    }
-
-    fn generate_type_info(typ: Option<Type<'a>>) -> String {
-        if let Some(typ) = typ {
-            format!("{:}", typ)
-        } else {
-            "None".to_string()
-        }
-    }
-}
-
-type M = Fixed<NodeIndex>;
-
-impl<'a, 'b> MonoidalReducerExpression<'a, M> for Dotify<'a, 'b> {
-    fn reduce_expression(&mut self, _input: &'a Expression<'a>, value: M) -> M {
-        // Bubble up value since `Expression` is an enum
-        value
-    }
-
-    fn reduce_array_access(&mut self, input: &'a ArrayAccessExpression<'a>, array: M, index: M) -> M {
-        let label = format!(
-            "Array Access Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        let Fixed(end_idx) = array;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("array".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = index;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("index".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_array_init(&mut self, input: &'a ArrayInitExpression<'a>, element: M) -> M {
-        let label = format!(
-            "Array Init Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        let Fixed(end_idx) = element;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("element".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_array_inline(&mut self, input: &'a ArrayInlineExpression<'a>, elements: Vec<M>) -> M {
-        let label = format!(
-            "Array Inline Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in elements.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("element_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_array_range_access(
-        &mut self,
-        input: &'a ArrayRangeAccessExpression<'a>,
-        array: M,
-        left: Option<M>,
-        right: Option<M>,
-    ) -> M {
-        let label = format!(
-            "Array Range Access Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        let Fixed(end_idx) = array;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("array".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(Fixed(end_idx)) = left {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("left".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(Fixed(end_idx)) = right {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("right".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_binary(&mut self, input: &'a BinaryExpression<'a>, left: M, right: M) -> M {
-        let label = format!(
-            "Binary Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        let Fixed(end_idx) = left;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("left".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = right;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("right".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_call(&mut self, input: &'a CallExpression<'a>, target: Option<M>, arguments: Vec<M>) -> M {
-        let label = format!(
-            "Call Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in arguments.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("argument_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(Fixed(end_idx)) = target {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("target".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            )
-        }
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_circuit_access(&mut self, input: &'a CircuitAccessExpression<'a>, target: Option<M>) -> M {
-        let label = format!(
-            "Circuit Access Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        if let Some(Fixed(end_idx)) = target {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("target".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_circuit_init(&mut self, input: &'a CircuitInitExpression<'a>, values: Vec<M>) -> M {
-        let label = format!(
-            "Circuit Init Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in values.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("value_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_ternary_expression(
-        &mut self,
-        input: &'a TernaryExpression<'a>,
-        condition: M,
-        if_true: M,
-        if_false: M,
-    ) -> M {
-        let label = format!(
-            "Ternary Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        let Fixed(end_idx) = condition;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("condition".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = if_true;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("if_true".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = if_false;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("if_false".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_cast_expression(&mut self, input: &'a CastExpression<'a>, inner: M) -> M {
-        let label = format!(
-            "Cast Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = inner;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("inner".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_lengthof_expression(&mut self, input: &'a LengthOfExpression<'a>, inner: M) -> M {
-        let label = format!(
-            "LengthOf Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = inner;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("inner".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_constant(&mut self, input: &'a Constant<'a>) -> M {
-        //TODO: Are there children to reduce?
-        //TODO: Implement Display for ConstValue
-        let label = format!(
-            "Constant\nNode ID: {:}\nType: {:}\nValue: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            "TODO".to_string(),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_tuple_access(&mut self, input: &'a TupleAccessExpression<'a>, tuple_ref: M) -> M {
-        let label = format!(
-            "Tuple Access Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = tuple_ref;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("tuple_ref".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_tuple_init(&mut self, input: &'a TupleInitExpression<'a>, values: Vec<M>) -> M {
-        let label = format!(
-            "Tuple Init Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in values.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("value_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_unary(&mut self, input: &'a UnaryExpression<'a>, inner: M) -> M {
-        let label = format!(
-            "Unary Expression\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = inner;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("inner".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_variable_ref(&mut self, input: &'a VariableRef<'a>) -> M {
-        //TODO: Are there children to visit here?
-        let label = format!(
-            "Variable Ref\nNode ID: {:}\nType: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_type_info(input.get_type()),
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        if let Some(parent) = input.parent.get() {
-            self.edges.push((
-                input.id,
-                parent.get_id(),
-                ("parent".to_string(), LabelType::Label),
-                Some(("red".to_string(), LabelType::Label)),
-            ));
-        }
-
-        Fixed(start_idx)
-    }
-}
-
-impl<'a, 'b> MonoidalReducerStatement<'a, M> for Dotify<'a, 'b> {
-    fn reduce_statement(&mut self, _input: &'a Statement<'a>, value: M) -> M {
-        // Just bubble up value as `Statement` is an enum
-        value
-    }
-
-    fn reduce_assign_access(&mut self, input: &AssignAccess<'a>, left: Option<M>, right: Option<M>) -> M {
-        // TODO: Monoidal reducer might need to be rewritter for this
-        let id = self.context.get_id();
-        let label = format!("Assign Access\nNode ID: {:}", id,);
-        let start_idx = self.add_or_get_node(id, label, LabelType::Esc);
-
-        if let Some(Fixed(end_idx)) = left {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("left".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        if let Some(Fixed(end_idx)) = right {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("right".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            )
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_assign(&mut self, input: &AssignStatement<'a>, accesses: Vec<M>, value: M) -> M {
-        let label = format!(
-            "Assign Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in accesses.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("access_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        let Fixed(end_idx) = value;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("value".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_block(&mut self, input: &BlockStatement<'a>, statements: Vec<M>) -> M {
-        let label = format!(
-            "Block Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in statements.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("statement_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_conditional_statement(
-        &mut self,
-        input: &ConditionalStatement<'a>,
-        condition: M,
-        if_true: M,
-        if_false: Option<M>,
-    ) -> M {
-        let label = format!(
-            "Console Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        let Fixed(end_idx) = condition;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("condition".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = if_true;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("if_true".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        if let Some(Fixed(end_idx)) = if_false {
-            self.add_edge(
-                start_idx,
-                end_idx,
-                ("condition".to_string(), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_formatted_string(&mut self, input: &ConsoleArgs<'a>, parameters: Vec<M>) -> M {
-        let label = format!("Console Args\nNode ID: {:}", input.id,);
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in parameters.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("parameter_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_console(&mut self, input: &ConsoleStatement<'a>, argument: M) -> M {
-        let label = format!(
-            "Console Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = argument;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("argument".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_definition(&mut self, input: &DefinitionStatement<'a>, value: M) -> M {
-        let label = format!(
-            "Definition Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = value;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("value".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_expression_statement(&mut self, input: &ExpressionStatement<'a>, expression: M) -> M {
-        let label = format!(
-            "Expression Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = expression;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("expression".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_iteration(&mut self, input: &IterationStatement<'a>, start: M, stop: M, body: M) -> M {
-        let label = format!(
-            "Iteration Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = start;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("start".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = stop;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("stop".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        let Fixed(end_idx) = body;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("body".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_return(&mut self, input: &ReturnStatement<'a>, value: M) -> M {
-        let label = format!(
-            "Return Statement\nNode ID: {:}\n\n{:}",
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = value;
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("value".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-}
-
-impl<'a, 'b> MonoidalReducerProgram<'a, M> for Dotify<'a, 'b> {
-    fn reduce_function(&mut self, input: &'a Function<'a>, body: M) -> M {
-        let label = format!(
-            "Function: {:}\nNode ID: {:}",
-            input.name.borrow().name.to_string(),
-            input.id
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-        let Fixed(end_idx) = body;
-
-        self.add_edge(
-            start_idx,
-            end_idx,
-            ("body".to_string(), LabelType::Label),
-            Some(("black".to_string(), LabelType::Label)),
-        );
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_circuit_member(&mut self, input: &CircuitMember<'a>, function: Option<M>) -> M {
-        //TODO: Need to figure out how to reduce types
-        //TODO: Might need to fix monoidal director
-        let start_idx = self.add_or_get_node(self.context.get_id(), "circuit_member".to_string(), LabelType::Label);
-        Fixed(start_idx)
-    }
-
-    fn reduce_circuit(&mut self, input: &'a Circuit<'a>, members: Vec<M>) -> M {
-        let label = format!(
-            "Circuit: {:}\nNode ID: {:}\n\n{:}",
-            input.name.borrow().name.to_string(),
-            input.id,
-            Dotify::generate_span_info(&input.span)
-        );
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in members.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("member_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        Fixed(start_idx)
-    }
-
-    fn reduce_program(&mut self, input: &Program, imported_modules: Vec<M>, functions: Vec<M>, circuits: Vec<M>) -> M {
-        let label = format!("Program: {:}\nNode ID: {:}", input.name, input.id);
-        let start_idx = self.add_or_get_node(input.id, label, LabelType::Esc);
-
-        for (i, Fixed(end_idx)) in functions.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("function_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        for (i, Fixed(end_idx)) in circuits.iter().enumerate() {
-            self.add_edge(
-                start_idx,
-                *end_idx,
-                (format!("circuit_{:}", i), LabelType::Label),
-                Some(("black".to_string(), LabelType::Label)),
-            );
-        }
-
-        self.add_remaining_edges();
-
-        Fixed(start_idx)
-    }
-}
-
-impl<'a, 'b> AsgPass<'a> for Dotify<'a, 'b> {
+impl<'a, 'b> AsgPass<'a> for Dotifier<'a, 'b> {
     type Input = (Program<'a>, &'b AsgContext<'a>, String, PathBuf);
     type Output = Result<Program<'a>>;
 
     fn do_pass((asg, ctx, id, path): Self::Input) -> Self::Output {
         let graph = DotGraph::new(id);
-        let mut director = MonoidalDirector::new(Dotify::new(graph, ctx));
+        let mut director = MonoidalDirector::new(Dotifier::new(graph, ctx));
         director.reduce_program(&asg);
 
         let mut file = File::create(path).unwrap();
